@@ -1,1 +1,6 @@
-<slot/>
+<script>
+	/** @type {{children?: import('svelte').Snippet}} */
+	let { children } = $props();
+</script>
+
+{@render children?.()}
