@@ -1,8 +1,13 @@
+<script>
+    /** @type {{children?: import('svelte').Snippet}} */
+    let { children } = $props();
+</script>
+
 <header>
     <h1>SMACK MAGIC</h1>
 </header>
 <main>
-    <slot/>
+    {@render children?.()}
 </main>
 <footer></footer>
 
